Set up repository and controller for post tags.

+++ Tabloid/Controllers/PostController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 using Tabloid.Data;
 using Tabloid.Models;
 using Tabloid.Repositories;
@@ -12,9 +13,11 @@ namespace Tabloid.Controllers
     public class PostController : ControllerBase
     {
         private readonly PostRepository _postRepository;
+        private readonly UserProfileRepository _userProfileRepository;
         public PostController(ApplicationDbContext context)
         {
             _postRepository = new PostRepository(context);
+            _userProfileRepository = new UserProfileRepository(context);
         }
 
         [HttpGet]
@@ -47,9 +50,30 @@ namespace Tabloid.Controllers
             return CreatedAtAction("Get", new { id = post.Id }, post);
         }
 
+        [HttpPost("addtag")]
+        public IActionResult Post(PostTag postTag)
+        {
+            var currentUserProfile = GetCurrentUserProfile();
+            var post = _postRepository.GetById(postTag.PostId);
+
+            if (currentUserProfile.Id != post.UserProfileId)
+            {
+                return Unauthorized();
+            }
+            _postRepository.InsertTag(postTag);
+            return CreatedAtAction("Get", new { id = postTag.Id }, postTag);
+        }
+
         [HttpPut("{id}")]
         public IActionResult Put(int id, Post post)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+
+            if (currentUserProfile.Id != post.UserProfileId)
+            {
+                return Unauthorized();
+            }
+
             if (id != post.Id)
             {
                 return BadRequest();
@@ -62,8 +86,22 @@ namespace Tabloid.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+            var post = _postRepository.GetById(id);
+
+            if (currentUserProfile.Id != post.UserProfileId)
+            {
+                return Unauthorized();
+            }
+
             _postRepository.Delete(id);
             return NoContent();
+        }
+
+        private UserProfile GetCurrentUserProfile()
+        {
+            var firebaseUserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            return _userProfileRepository.GetByFirebaseUserId(firebaseUserId);
         }
     }
 }
